+++ R/equal.R
@@ -66,11 +66,7 @@ vec_proxy_equal.default <- function(x, ...) {
 vec_equal <- function(x, y, na_equal = FALSE, .ptype = NULL) {
   vec_assert(na_equal, ptype = logical(), size = 1L)
   args <- vec_recycle_common(x, y)
-  args <- vec_cast_common_params(
-    !!!args,
-    .to = .ptype,
-    .df_fallback = DF_FALLBACK_quiet
-  )
+  args <- vec_cast_common_params(!!!args, .to = .ptype)
   .Call(vctrs_equal, args[[1]], args[[2]], na_equal)
 }