+++ R/subscript-loc.R
@@ -259,29 +259,24 @@ vec_as_location2_result <- function(i,
     )))
   }
 
-  # FIXME: Use result approach in internal implementation?
   err <- NULL
   i <- tryCatch(
     vec_as_location(i, n, names = names, arg = arg, call = call),
-    vctrs_error_subscript_type = function(err) {
+    vctrs_error_subscript = function(err) {
+      err[["subscript_scalar"]] <- TRUE
       err <<- err
       i
     }
   )
+  if (!is_null(err)) {
+    return(result(err = err))
+  }
 
   if (neg) {
     i <- -i
   }
 
-  if (is_null(err)) {
-    result(i)
-  } else {
-    result(err = new_error_location2_type(
-      i = i,
-      subscript_arg = arg,
-      call = call
-    ))
-  }
+  result(i)
 }
 
 
@@ -354,19 +349,19 @@ new_error_location2_type <- function(i,
 cnd_bullets_location2_need_scalar <- function(cnd, ...) {
   cnd$subscript_arg <- append_arg("Subscript", cnd$subscript_arg)
   format_error_bullets(c(
-    x = glue::glue_data(cnd, "{subscript_arg} has size {length(i)} but must be size 1.")
+    x = glue::glue_data(cnd, "{subscript_arg} must be size 1, not {length(i)}.")
   ))
 }
 cnd_bullets_location2_need_present <- function(cnd, ...) {
   cnd$subscript_arg <- append_arg("Subscript", cnd$subscript_arg)
   format_error_bullets(c(
-    x = glue::glue_data(cnd, "{subscript_arg} can't be `NA`.")
+    x = glue::glue_data(cnd, "{subscript_arg} must be a location, not {obj_type_friendly(i)}.")
   ))
 }
 cnd_bullets_location2_need_positive <- function(cnd, ...) {
   cnd$subscript_arg <- append_arg("Subscript", cnd$subscript_arg)
   format_error_bullets(c(
-    x = glue::glue_data(cnd, "{subscript_arg} has value {i} but must be a positive location.")
+    x = glue::glue_data(cnd, "{subscript_arg} must be a positive location, not {i}.")
   ))
 }
 
@@ -472,11 +467,10 @@ stop_indicator_size <- function(i, n, ..., call = caller_env()) {
   ))
 }
 cnd_body_vctrs_error_indicator_size <- function(cnd, ...) {
-  cnd$subscript_arg <- append_arg("subscript", cnd$subscript_arg)
+  cnd$subscript_arg <- append_arg("Logical subscript", cnd$subscript_arg)
   glue_data_bullets(
     cnd,
-    i = "Logical subscripts must match the size of the indexed input.",
-    x = "Input has size {n} but {subscript_arg} has size {vec_size(i)}."
+    x = "{subscript_arg} must be size 1 or {n}, not {vec_size(i)}."
   )
 }
 
@@ -611,6 +605,6 @@ cnd_body_vctrs_error_subscript_oob_non_consecutive <- function(cnd, ...) {
 
 cnd_subscript_oob_non_consecutive <- function(cnd) {
   out <- cnd$subscript_oob_non_consecutive %||% FALSE
-  stopifnot(is_bool(out))
+  check_bool(out)
   out
 }