+++ R/type2.R
@@ -36,6 +36,14 @@ vec_ptype2 <- function(x,
                        call = caller_env()) {
   if (!missing(...)) {
     check_ptype2_dots_empty(...)
+    return(vec_ptype2_opts(
+      x,
+      y,
+      opts = match_fallback_opts(...),
+      x_arg = x_arg,
+      y_arg = y_arg,
+      call = call
+    ))
   }
   return(.Call(ffi_ptype2, x, y, environment()))
   UseMethod("vec_ptype2")
@@ -117,10 +125,15 @@ vec_default_ptype2 <- function(x,
 
   opts <- match_fallback_opts(...)
 
-  # If both data frames, first find common type of columns before the
-  # same-type fallback
-  if (df_needs_normalisation(x, y, opts)) {
-    out <- vec_ptype2_df_fallback_normalise(
+  if (opts$s3_fallback && can_fall_back_2(x, y)) {
+    common <- common_class_suffix(x, y)
+    if (length(common)) {
+      return(new_common_class_fallback(x, common))
+    }
+  }
+
+  if (is.data.frame(x) && is.data.frame(y)) {
+    out <- vec_ptype2_df_fallback(
       x,
       y,
       opts,
@@ -128,42 +141,16 @@ vec_default_ptype2 <- function(x,
       y_arg = y_arg,
       call = call
     )
-    x <- out$x
-    y <- out$y
-  }
 
-  if (opts$s3_fallback && can_fall_back(x, y)) {
-    common <- common_class_suffix(x, y)
-    if (length(common)) {
-      return(new_common_class_fallback(x, common))
+    if (identical(non_df_attrib(x), non_df_attrib(y))) {
+      attributes(out) <- c(df_attrib(out), non_df_attrib(x))
     }
+
+    return(out)
   }
 
   if (is_same_type(x, y)) {
     return(vec_ptype(x, x_arg = x_arg))
-  }
-
-  if (has_df_fallback(opts$df_fallback)) {
-    if (is_df_subclass(x) && is.data.frame(y)) {
-      return(vec_ptype2_df_fallback(
-        x,
-        y,
-        opts,
-        x_arg = x_arg,
-        y_arg = y_arg,
-        call = call
-      ))
-    }
-    if (is_df_subclass(y) && is.data.frame(x)) {
-      return(vec_ptype2_df_fallback(
-        x,
-        y,
-        opts,
-        x_arg = x_arg,
-        y_arg = y_arg,
-        call = call
-      ))
-    }
   }
 
   # The from-dispatch parameter is set only when called from our S3
@@ -206,28 +193,52 @@ vec_incompatible_ptype2 <- function(x,
 # We can't check for a proxy or ptype2 method to determine whether a
 # class is foreign, because we implement these generics for many base
 # classes and we still need to allow base fallbacks with subclasses.
-can_fall_back <- function(x, y) {
-  # Work around bug with hard-coded `tsp` attribute in Rf_setAttrib()
-  if (inherits(x, "ts") || inherits(y, "ts")) {
-    return(FALSE)
-  }
-
-  if (is.data.frame(x) || is.data.frame(y)) {
-    return(FALSE)
-  }
-
+can_fall_back_2 <- function(x, y) {
   if (!identical(typeof(x), typeof(y))) {
     return(FALSE)
   }
 
-  # Suboptimal: Prevent bad interaction with proxy-assign
-  has_no_proxy(x) && has_no_proxy(y)
-}
-has_no_proxy <- function(x) {
-  if (inherits(x, "vctrs:::common_class_fallback")) {
-    return(TRUE)
+  if (!can_fall_back(x) || !can_fall_back(y)) {
+    return(FALSE)
   }
 
+  TRUE
+}
+
+can_fall_back <- function(x) {
+  UseMethod("can_fall_back")
+}
+
+#' @export
+can_fall_back.vctrs_vctr <- function(x) {
+  # Work aronud bad interaction when `c()` method calls back into `vec_c()`
+  FALSE
+}
+#' @export
+can_fall_back.ts <- function(x) {
+  # Work around bug with hard-coded `tsp` attribute in Rf_setAttrib()
+  FALSE
+}
+#' @export
+can_fall_back.data.frame <- function(x) {
+  # The `c()` fallback is only for 1D vectors
+  FALSE
+}
+
+#' @export
+`can_fall_back.vctrs:::common_class_fallback` <- function(x) {
+  TRUE
+}
+
+#' @export
+can_fall_back.default <- function(x) {
+  # Don't all back for classes that directly implement a proxy.
+  #
+  # NOTE: That's suboptimal. For instance this forces us to override
+  # `can_fall_back()` for `vctrs_vctr` to avoid recursing into
+  # `vec_c()` through `c()`. Maybe we want to avoid falling back for
+  # any vector that inherits a `vec_proxy()` method implemented
+  # _outside_ of vctrs, i.e. not for a base class?
   is_null(s3_get_method(class(x)[[1]], "vec_proxy", ns = "vctrs"))
 }
 
@@ -259,15 +270,11 @@ fallback_class <- function(x) {
 
 check_ptype2_dots_empty <- function(...,
                                     `vctrs:::from_dispatch`,
-                                    `vctrs:::df_fallback`,
                                     `vctrs:::s3_fallback`) {
   check_dots_empty0(...)
 }
-match_fallback_opts <- function(...,
-                                `vctrs:::df_fallback` = NULL,
-                                `vctrs:::s3_fallback` = NULL) {
+match_fallback_opts <- function(..., `vctrs:::s3_fallback` = NULL) {
   fallback_opts(
-    df_fallback = `vctrs:::df_fallback`,
     s3_fallback = `vctrs:::s3_fallback`
   )
 }
@@ -275,18 +282,15 @@ match_from_dispatch <- function(..., `vctrs:::from_dispatch` = FALSE) {
   `vctrs:::from_dispatch`
 }
 
-fallback_opts <- function(df_fallback = NULL,
-                          s3_fallback = NULL) {
+fallback_opts <- function(s3_fallback = NULL) {
   # Order is important for the C side
   list(
-    df_fallback = df_fallback %||% df_fallback_default(),
     s3_fallback = s3_fallback %||% s3_fallback_default()
   )
 }
 
 full_fallback_opts <- function() {
   fallback_opts(
-    df_fallback = DF_FALLBACK_quiet,
     s3_fallback = S3_FALLBACK_true
   )
 }
@@ -303,13 +307,11 @@ vec_ptype2_opts <- function(x,
 vec_ptype2_params <- function(x,
                               y,
                               ...,
-                              df_fallback = NULL,
                               s3_fallback = NULL,
                               x_arg = "",
                               y_arg = "",
                               call = caller_env()) {
   opts <- fallback_opts(
-    df_fallback = df_fallback,
     s3_fallback = s3_fallback
   )
   vec_ptype2_opts(
@@ -329,7 +331,6 @@ vec_ptype2_no_fallback <- function(x,
                                    y_arg = "",
                                    call = caller_env()) {
   opts <- fallback_opts(
-    df_fallback = DF_FALLBACK_none,
     s3_fallback = S3_FALLBACK_false
   )
   vec_ptype2_opts(
@@ -344,35 +345,10 @@ vec_ptype2_no_fallback <- function(x,
   )
 }
 
-
-# Kept in sync with ptype2.h
-df_fallback_default <- function() 0L
-DF_FALLBACK_warn_maybe <- 0L
-DF_FALLBACK_warn <- 1L
-DF_FALLBACK_none <- 2L
-DF_FALLBACK_quiet <- 3L
-
 s3_fallback_default <- function() 0L
 S3_FALLBACK_false <- 0L
 S3_FALLBACK_true <- 1L
 
-
-has_df_fallback <- function(df_fallback) {
-  df_fallback != DF_FALLBACK_none
-}
-needs_fallback_warning <- function(df_fallback) {
-  if (df_fallback == DF_FALLBACK_warn_maybe) {
-    is_true(peek_option("vctrs:::warn_on_fallback"))
-  } else {
-    df_fallback == DF_FALLBACK_warn
-  }
-}
-with_fallback_warning <- function(expr) {
-  with_options(expr, `vctrs:::warn_on_fallback` = TRUE)
-}
-with_fallback_quiet <- function(expr) {
-  with_options(expr, `vctrs:::warn_on_fallback` = FALSE)
-}
 
 vec_typeof2 <- function(x, y) {
   .Call(ffi_typeof2, x, y)