+++ src/c-unchop.c
@@ -35,7 +35,6 @@ r_obj* list_unchop(r_obj* xs,
 
   ptype = KEEP(vec_ptype_common_params(xs,
                                        ptype,
-                                       DF_FALLBACK_DEFAULT,
                                        S3_FALLBACK_true,
                                        p_error_arg,
                                        error_call));
@@ -167,7 +166,6 @@ r_obj* list_unchop(r_obj* xs,
   if (is_data_frame(proxy)) {
     df_c_fallback(proxy, ptype, xs, out_size, name_spec, name_repair, error_call);
   }
-
   r_obj* out = KEEP(vec_restore_recurse(proxy, ptype, VCTRS_OWNED_true));
 
   if (out_names != r_null) {