+++ tests/testthat/helper-vctrs.R
@@ -3,17 +3,6 @@ testthat_import_from <- function(ns, names, env = caller_env()) {
   import_from(ns, names, env = env)
 }
 
-vec_ptype2_fallback <- function(x, y, ...) {
-  vec_ptype2_params(x, y, ..., df_fallback = DF_FALLBACK_warn)
-}
-vec_ptype_common_df_fallback <- function(..., .ptype = NULL) {
-  vec_ptype_common_params(
-    ...,
-    .ptype = .ptype,
-    .df_fallback = DF_FALLBACK_warn
-  )
-}
-
 shaped_int <- function(...) {
   array(NA_integer_, c(...))
 }