+++ tests/testthat/test-cast.R
@@ -93,11 +93,10 @@ test_that("unspecified can be cast to shaped vectors", {
 })
 
 test_that("vec_cast() only falls back when casting to base type", {
-  expect_incompatible_df_cast(vec_cast(foobar(mtcars), mtcars), mtcars)
-  expect_error(
-    vec_cast(mtcars, foobar(mtcars)),
-    class = "vctrs_error_incompatible_type"
-  )
+  expect_equal(vec_cast(foobar(mtcars), mtcars), mtcars)
+  expect_snapshot({
+    (expect_error(vec_cast(mtcars, foobar(mtcars))))
+  })
 })
 
 test_that("vec_cast() only attempts to fall back if `to` is a data frame (#1568)", {
@@ -265,4 +264,23 @@ test_that("can call `vec_cast()` from C (#1666)", {
   y <- array(2, dim = c(2, 2))
 
   expect_equal(fn(x, y), vec_cast(x, y))
+})
+
+test_that("df-fallback for cast is not sensitive to attributes order", {
+  x <- structure(
+    list(col = ""),
+    class = c("vctrs_foobar", "tbl_df", "tbl", "data.frame"),
+    row.names = c(NA, -1L),
+    foo = "foo",
+    bar = "bar"
+  )
+  ptype <- structure(
+    list(col = character(0)),
+    foo = "foo",
+    bar = "bar",
+    row.names = integer(0),
+    class = c("vctrs_foobar", "tbl_df", "tbl", "data.frame")
+  )
+
+  expect_identical(vec_cast(x, ptype), x)
 })